move serializer to separate file
