+++ src/middleware/mod.rs
@@ -18,6 +18,7 @@ mod error;
 mod operation;
 mod pod_deserialization;
 pub mod serialization;
+pub mod serializer;
 mod statement;
 use std::{any::Any, collections::HashMap, fmt};
